Moved r_sexp methods to single header
