+++ inst/include/cppally/sugar/r_combine.h
@@ -6,6 +6,7 @@
 #include <cppally/sugar/r_subset.h>
 #include <cppally/sugar/r_rep.h>
 #include <cppally/sugar/r_make_vec.h>
+#include <cppally/sugar/r_sexp_methods.h>
 
 namespace cppally {
