+++ inst/include/cppally/sugar/r_sort.h
@@ -410,12 +410,7 @@ inline r_vec<r_int> order(const r_df& x, bool preserve_ties = true) {
 }
 
 
-inline r_vec<r_int> order(const r_sexp& x, bool preserve_ties) {
-    return CPPALLY_VIEW_AND_APPLY(
-        x, /*return_type = */ r_vec<r_int>, /*fn = */ order, 
-        /*rest of args = */ preserve_ties
-    );
-}
+inline r_vec<r_int> order(const r_sexp& x, bool preserve_ties);
 
 // Is x in a sorted order? i.e is x increasing but not necessarily monotonically?
 // To retrieve a bool result, use the `is_true` member function
